t/tree-test-phrasing.dat

#data
</sarcasm>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<sarcasm>aaa</sarcasm>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <sarcasm>
|       "aaa"

#data
<!DOCTYPE HTML><rt></body>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <rt>

#data
<!DOCTYPE HTML><rp></body>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <rp>

#data
<!DOCTYPE HTML><span><rp></body>
#errors
span not closed
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <span>
|       <rp>

#data
<!DOCTYPE HTML><rp><span></body>
#errors
span not closed
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <rp>
|       <span>

#data
<!DOCTYPE HTML>
<cite><b><cite><code><cite><code><cite><code><div>X</b>TEST
#errors
aaa
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <cite>
|       <b>
|         <cite>
|           <code>
|             <cite>
|               <code>
|                 <cite>
|                   <code>
|       <code>
|         <code>
|           <code>
|             <div>
|               <b>
|                 "X"
|               "TEST"
